feat: Update tools.py module with new functions and improvements
This commit updates the tools.py module with new
functions and improvements, including:

* Added vector_line and vector_line_eqn functions for
calculating vector differences and equations for
line segments.
* Added angle_between_vectors function for calculating
the angle between two vectors.
* Improved create_vectors function to accept
variable number of arguments.
* Added create_3d_components function for creating
3D components for given symbols.
* Added print_aligned_latex_equations function for
printing aligned LaTeX equations using IPython display.
* Added unit tests for new functions using the
doctest module.
These changes provide a more comprehensive set of
tools for working with vectors and reference frames
in three-dimensional space, which can be useful
in a variety of applications, including physics,
engineering, and computer graphics.

By using the Sympy physics vector module, the functions
in this module can be used to perform symbolic
vector operations, which can be useful for
analytical calculations and derivations.

Changes to be committed:
	deleted:    LICENSE
	new file:   environment.yml
	new file:   setup.py
	renamed:    sympytools/__init.__py -> sympytools/__init__.py
	modified:   sympytools/chainrule.py
	modified:   sympytools/tools.py